user_prac/app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-login==1.10.2
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
